+++ src/plugin/gui/controls/gui_filter_visualiser.cpp
@@ -128,8 +128,6 @@ void DoFilterVisualizer(GuiState& g, u8 layer_index, Rect viewport_r, bool greye
     imgui.PushId(push_id);
     DEFER { imgui.PopId(); };
 
-    filter_display::DrawBackground(imgui, viewport_r);
-
     auto const cutoff_param = params.DescribedValue(layer_index, LayerParamIndex::FilterCutoff);
     auto const reso_param = params.DescribedValue(layer_index, LayerParamIndex::FilterResonance);
     auto const type_param = params.DescribedValue(layer_index, LayerParamIndex::FilterType);
@@ -137,9 +135,10 @@ void DoFilterVisualizer(GuiState& g, u8 layer_index, Rect viewport_r, bool greye
     auto const cutoff_index = ParamIndexFromLayerParamIndex(layer_index, LayerParamIndex::FilterCutoff);
     auto const reso_index = ParamIndexFromLayerParamIndex(layer_index, LayerParamIndex::FilterResonance);
 
+    filter_display::DrawBackground(imgui, viewport_r, cutoff_param.info);
+
     auto const node_pos_viewport = [&] {
-        auto const cutoff_hz = cutoff_param.info.ProjectValue(cutoff_param.LinearValue());
-        return f32x2 {viewport_r.x + (filter_display::HzToX01(cutoff_hz) * viewport_r.w),
+        return f32x2 {viewport_r.x + (cutoff_param.LinearValue() * viewport_r.w),
                       filter_display::DbToY(0.0f, viewport_r)};
     };
 
@@ -173,9 +172,7 @@ void DoFilterVisualizer(GuiState& g, u8 layer_index, Rect viewport_r, bool greye
 
         auto const cursor_x = GuiIo().in.cursor_pos.x - rel_click_pos.x;
         auto const x_clamped = Clamp(cursor_x, min_x, max_x);
-        auto const x_t = MapTo01(x_clamped, min_x, max_x);
-        auto const target_hz = filter_display::X01ToHz(x_t);
-        auto const new_cutoff_linear = cutoff_param.info.LineariseValue(target_hz, true).ValueOr(0.0f);
+        auto const new_cutoff_linear = MapTo01(x_clamped, min_x, max_x);
 
         SetParameterValue(engine.processor, cutoff_index, new_cutoff_linear, {});
     }
@@ -222,6 +219,7 @@ void DoFilterVisualizer(GuiState& g, u8 layer_index, Rect viewport_r, bool greye
                                               res_adj,
                                               hz);
             },
+            cutoff_param.info,
             greyed_out);
     }
 
@@ -335,16 +333,15 @@ void DoEffectFilterVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) {
     imgui.PushId(SourceLocationHash());
     DEFER { imgui.PopId(); };
 
-    filter_display::DrawBackground(imgui, viewport_r);
-
     auto const cutoff_param = params.DescribedValue(ParamIndex::FilterCutoff);
     auto const reso_param = params.DescribedValue(ParamIndex::FilterResonance);
     auto const gain_param = params.DescribedValue(ParamIndex::FilterGain);
     auto const type_param = params.DescribedValue(ParamIndex::FilterType);
 
+    filter_display::DrawBackground(imgui, viewport_r, cutoff_param.info);
+
     auto const uses_gain = [&] {
-        return param_values::EffectFilterTypeUsesGain(
-            type_param.IntValue<param_values::EffectFilterType>());
+        return param_values::EffectFilterTypeUsesGain(type_param.IntValue<param_values::EffectFilterType>());
     };
 
     auto const num_stages = [&] {
@@ -352,9 +349,8 @@ void DoEffectFilterVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) {
     };
 
     auto const node_pos_viewport = [&] {
-        auto const cutoff_hz = cutoff_param.info.ProjectValue(cutoff_param.LinearValue());
         return f32x2 {
-            viewport_r.x + (filter_display::HzToX01(cutoff_hz) * viewport_r.w),
+            viewport_r.x + (cutoff_param.LinearValue() * viewport_r.w),
             uses_gain()
                 ? filter_display::DbToY(
                       Clamp(gain_param.ProjectedValue(), filter_display::k_min_db, filter_display::k_max_db),
@@ -395,9 +391,7 @@ void DoEffectFilterVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) {
 
         auto const cursor = GuiIo().in.cursor_pos - rel_click_pos;
         auto const x_clamped = Clamp(cursor.x, min_x, max_x);
-        auto const x_t = MapTo01(x_clamped, min_x, max_x);
-        auto const target_hz = filter_display::X01ToHz(x_t);
-        auto const new_cutoff_linear = cutoff_param.info.LineariseValue(target_hz, true).ValueOr(0.0f);
+        auto const new_cutoff_linear = MapTo01(x_clamped, min_x, max_x);
         SetParameterValue(engine.processor, ParamIndex::FilterCutoff, new_cutoff_linear, {});
 
         if (uses_gain()) {
@@ -458,6 +452,7 @@ void DoEffectFilterVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) {
         [&](f32 hz) {
             return (f32)stages * rbj_filter::MagnitudeDb(coeffs, hz, filter_display::k_sample_rate);
         },
+        cutoff_param.info,
         greyed_out);
 
     DescribedParamValue const* params_arr[] = {&cutoff_param, &reso_param, &gain_param};
@@ -538,7 +533,8 @@ void DoReverbPreFilterVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) 
     imgui.PushId(SourceLocationHash());
     DEFER { imgui.PopId(); };
 
-    filter_display::DrawBackground(imgui, viewport_r);
+    auto const& freq_info = k_param_descriptors[ToInt(ParamIndex::FilterCutoff)];
+    filter_display::DrawBackground(imgui, viewport_r, freq_info);
 
     auto const lp_param = params.DescribedValue(ParamIndex::ReverbPreLowPassCutoff);
     auto const hp_param = params.DescribedValue(ParamIndex::ReverbPreHighPassCutoff);
@@ -557,7 +553,8 @@ void DoReverbPreFilterVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) 
 
     auto const node_pos_viewport = [&](Grabber const& gr) {
         auto const cutoff_hz = SemitonesToHz(gr.param.LinearValue());
-        return f32x2 {viewport_r.x + (filter_display::HzToX01(cutoff_hz) * viewport_r.w), handle_y};
+        auto const x_lin = freq_info.LineariseValue(cutoff_hz, true).ValueOr(0.0f);
+        return f32x2 {viewport_r.x + (x_lin * viewport_r.w), handle_y};
     };
 
     struct GrabberFrame {
@@ -603,7 +600,7 @@ void DoReverbPreFilterVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) 
             auto const cursor_x = GuiIo().in.cursor_pos.x - rel_click_pos.x;
             auto const x_clamped = Clamp(cursor_x, min_x, max_x);
             auto const x_t = MapTo01(x_clamped, min_x, max_x);
-            auto const target_hz = filter_display::X01ToHz(x_t);
+            auto const target_hz = freq_info.ProjectValue(x_t);
             auto const semitones = Clamp(FrequencyToMidiNote(target_hz),
                                          gr.param.info.linear_range.min,
                                          gr.param.info.linear_range.max);
@@ -623,6 +620,7 @@ void DoReverbPreFilterVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) 
         imgui,
         viewport_r,
         [&](f32 hz) { return LpMagDb(hz, lp_fc_hz) + HpMagDb(hz, hp_fc_hz); },
+        freq_info,
         greyed_out);
 
     for (auto const i : Range(ArraySize(grabbers))) {
@@ -668,7 +666,8 @@ void DoReverbPostShelfVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) 
     imgui.PushId(SourceLocationHash());
     DEFER { imgui.PopId(); };
 
-    filter_display::DrawBackground(imgui, viewport_r);
+    auto const& freq_info = k_param_descriptors[ToInt(ParamIndex::FilterCutoff)];
+    filter_display::DrawBackground(imgui, viewport_r, freq_info);
 
     auto const lo_cut_param = params.DescribedValue(ParamIndex::ReverbLowShelfCutoff);
     auto const lo_gain_param = params.DescribedValue(ParamIndex::ReverbLowShelfGain);
@@ -689,8 +688,9 @@ void DoReverbPostShelfVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) 
 
     auto const node_pos_viewport = [&](Shelf const& sh) {
         auto const cutoff_hz = SemitonesToHz(sh.cutoff_param.LinearValue());
+        auto const x_lin = freq_info.LineariseValue(cutoff_hz, true).ValueOr(0.0f);
         return f32x2 {
-            viewport_r.x + (filter_display::HzToX01(cutoff_hz) * viewport_r.w),
+            viewport_r.x + (x_lin * viewport_r.w),
             filter_display::DbToY(
                 Clamp(sh.gain_param.ProjectedValue(), filter_display::k_min_db, filter_display::k_max_db),
                 viewport_r),
@@ -745,7 +745,7 @@ void DoReverbPostShelfVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) 
 
             auto const x_clamped = Clamp(cursor.x, min_x, max_x);
             auto const x_t = MapTo01(x_clamped, min_x, max_x);
-            auto const target_hz = filter_display::X01ToHz(x_t);
+            auto const target_hz = freq_info.ProjectValue(x_t);
             auto const semitones = Clamp(FrequencyToMidiNote(target_hz),
                                          sh.cutoff_param.info.linear_range.min,
                                          sh.cutoff_param.info.linear_range.max);
@@ -779,6 +779,7 @@ void DoReverbPostShelfVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) 
         [&](f32 hz) {
             return LowShelfMagDb(hz, lo_fc_hz, lo_gain_db) + HighShelfMagDb(hz, hi_fc_hz, hi_gain_db);
         },
+        freq_info,
         greyed_out);
 
     for (auto const i : Range(ArraySize(shelves))) {
@@ -834,14 +835,16 @@ void DoDelayFilterVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) {
     imgui.PushId(SourceLocationHash());
     DEFER { imgui.PopId(); };
 
-    filter_display::DrawBackground(imgui, viewport_r);
+    auto const& freq_info = k_param_descriptors[ToInt(ParamIndex::FilterCutoff)];
+    filter_display::DrawBackground(imgui, viewport_r, freq_info);
 
     auto const cutoff_param = params.DescribedValue(ParamIndex::DelayFilterCutoffSemitones);
     auto const spread_param = params.DescribedValue(ParamIndex::DelayFilterSpread);
 
     auto const node_pos_viewport = [&] {
         auto const cutoff_hz = SemitonesToHz(cutoff_param.LinearValue());
-        return f32x2 {viewport_r.x + (filter_display::HzToX01(cutoff_hz) * viewport_r.w),
+        auto const x_lin = freq_info.LineariseValue(cutoff_hz, true).ValueOr(0.0f);
+        return f32x2 {viewport_r.x + (x_lin * viewport_r.w),
                       viewport_r.Bottom() - (spread_param.LinearValue() * viewport_r.h)};
     };
 
@@ -880,7 +883,7 @@ void DoDelayFilterVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) {
 
         auto const x_clamped = Clamp(cursor.x, min_x, max_x);
         auto const x_t = MapTo01(x_clamped, min_x, max_x);
-        auto const target_hz = filter_display::X01ToHz(x_t);
+        auto const target_hz = freq_info.ProjectValue(x_t);
         auto const semitones = Clamp(FrequencyToMidiNote(target_hz),
                                      cutoff_param.info.linear_range.min,
                                      cutoff_param.info.linear_range.max);
@@ -910,6 +913,7 @@ void DoDelayFilterVisualizer(GuiState& g, Rect viewport_r, bool greyed_out) {
         imgui,
         viewport_r,
         [&](f32 hz) { return LpMagDb(hz, lp_fc_hz) + HpMagDb(hz, hp_fc_hz); },
+        freq_info,
         greyed_out);
 
     DescribedParamValue const* params_arr[] = {&cutoff_param, &spread_param};